# internal/schema/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAgentPlatform API
  version: 1.1.0
  description: |
    REST API for OpenAgentPlatform (OAP) — an open-source, agent-first
    Remote Monitoring and Management (RMM) platform.

    ## Authentication
    All `/api/v1/*` endpoints require a bearer token (OIDC ID token or
    session JWT). Auth endpoints under `/auth/*` use OIDC Authorization
    Code Flow with PKCE.

    ## Versioning
    The API is versioned via URL path (`/api/v1`). Breaking changes
    increment the major version; additive changes are backward compatible.

    ## Rate Limiting
    Default rate limit is 1000 requests/min per token. Burst to 2000.
  contact:
    name: OpenAgentPlatform
    url: https://openagentplatform.io
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.openagentplatform.io
    description: Production
  - url: https://staging-api.openagentplatform.io
    description: Staging
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/scripts:
    get:
      tags: [Scripts]
      summary: List scripts
      operationId: listScripts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: language
          in: query
          schema:
            type: string
            enum: [bash, powershell, python, zsh]
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated script list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedScript"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Scripts]
      summary: Create script
      operationId: createScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptCreate"
      responses:
        "201":
          description: Script created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/scripts/{id}:
    get:
      tags: [Scripts]
      summary: Get script
      operationId: getScript
      parameters:
        - $ref: "#/components/parameters/ScriptIdPath"
      responses:
        "200":
          description: Script details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/scripts/{id}/run:
    post:
      tags: [Scripts]
      summary: Run script on agents
      operationId: runScript
      parameters:
        - $ref: "#/components/parameters/ScriptIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptRunRequest"
      responses:
        "202":
          description: Run enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/scripts/{id}/runs:
    get:
      tags: [Scripts]
      summary: List script runs
      description: Returns historical execution runs of this script.
      operationId: listScriptRuns
      parameters:
        - $ref: "#/components/parameters/ScriptIdPath"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated script runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedScriptRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # =========================================================
  # Audit (Phase 0)
  # =========================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC ID token or session JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      description: 1-based page index
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page (max 200)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ScriptIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized — missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden — token lacks required role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: invalid_request
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Correlation ID for log lookup
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    Script:
      type: object
      required: [id, name, language, body, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        language:
          type: string
          enum: [bash, powershell, python, zsh]
        body:
          type: string
        default_args:
          type: array
          items:
            type: string
        default_timeout_seconds:
          type: integer
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ScriptCreate:
      type: object
      required: [name, language, body]
      properties:
        name:
          type: string
        description:
          type: string
        language:
          type: string
          enum: [bash, powershell, python, zsh]
        body:
          type: string
        default_args:
          type: array
          items:
            type: string
        default_timeout_seconds:
          type: integer
        tags:
          type: array
          items:
            type: string
    ScriptRun:
      type: object
      required: [id, script_id, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        script_id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed, timeout]
        args:
          type: array
          items:
            type: string
        exit_code:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    ScriptRunRequest:
      type: object
      required: [agent_ids]
      properties:
        agent_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        args:
          type: array
          items:
            type: string
        timeout_seconds:
          type: integer
        env:
          type: object
          additionalProperties:
            type: string
...
    PaginatedScript:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Script"
        page:
          $ref: "#/components/schemas/PageMeta"
    PaginatedScriptRun:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScriptRun"
        page:
          $ref: "#/components/schemas/PageMeta"
...